@@ -56,9 +56,9 @@
     "@babel/preset-typescript": "^7.16.7",
     "@babel/register": "^7.17.7",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.5",
-    "@types/node": "^17.0.24",
-    "@types/react": "^17",
-    "@types/react-dom": "^17",
+    "@types/node": "^17.0.25",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "@types/webpack": "^5.28.0",
     "babel-loader": "^8.2.4",
     "concurrently": "^7.1.0",
@@ -66,11 +66,11 @@
     "css-loader": "^6.7.1",
     "electron": "^18",
     "electron-builder": "^23.0.3",
-    "react": "^17",
-    "react-dom": "^17",
+    "react": "^18",
+    "react-dom": "^18",
     "react-refresh": "^0.12.0",
     "rimraf": "^3.0.2",
-    "sass": "^1.50.0",
+    "sass": "^1.50.1",
     "sass-loader": "^12.6.0",
     "style-loader": "^3.3.1",
     "ts-node": "^10.7.0",
